# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  approval-engine-service_internal_domain.ApprovalAssignment:
    properties:
      acted_at:
        type: string
      comment:
        type: string
      id:
        type: string
      request_id:
        type: string
      status:
        type: string
      step_id:
        type: string
      user_id:
        type: string
      user_name:
        type: string
      user_position:
        type: string
    type: object
  approval-engine-service_internal_domain.ApprovalRequest:
    properties:
      app_id:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      current_step_order:
        type: integer
      definition_id:
        type: string
      doc_type:
        type: string
      id:
        type: string
      payload:
        additionalProperties: {}
        type: object
      requester_id:
        type: string
      resource_id:
        type: string
      status:
        type: string
      steps:
        items:
          $ref: '#/definitions/approval-engine-service_internal_domain.ApprovalStep'
        type: array
    type: object
  approval-engine-service_internal_domain.ApprovalStep:
    properties:
      activated_at:
        type: string
      approval_mode:
        type: string
      assignments:
        items:
          $ref: '#/definitions/approval-engine-service_internal_domain.ApprovalAssignment'
        type: array
      completed_at:
        type: string
      id:
        type: string
      name:
        type: string
      request_id:
        type: string
      status:
        type: string
      step_order:
        type: integer
    type: object
  approval-engine-service_internal_domain.Condition:
    properties:
      field:
        type: string
      op:
        description: eq, ne, gt, gte, lt, lte, in
        type: string
      value: {}
    type: object
...
  approval-engine-service_internal_domain.ResolverRule:
    properties:
      department:
        description: 'role: explicit = cross-function hop'
        type: string
      level:
        description: superior
        type: integer
      path:
        description: 'field: key inside the request payload'
        type: string
      position:
        description: role
        type: string
      scope:
        description: 'role: "same_department"'
        type: string
      type:
        type: string
      user_id:
        description: static
        type: string
    type: object
  approval-engine-service_internal_domain.WorkflowDefinition:
    properties:
      app_id:
        type: string
      created_at:
        type: string
      doc_type:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      steps:
        items:
          $ref: '#/definitions/approval-engine-service_internal_domain.WorkflowStep'
        type: array
      version:
        type: integer
    type: object
  approval-engine-service_internal_domain.WorkflowStep:
    properties:
      approval_mode:
        type: string
      condition:
        $ref: '#/definitions/approval-engine-service_internal_domain.Condition'
      conditions:
        items:
          $ref: '#/definitions/approval-engine-service_internal_domain.Condition'
        type: array
      definition_id:
        type: string
      id:
        type: string
      logic:
        description: LogicAll (default) or LogicAny; only used with Conditions
        type: string
      name:
        type: string
      on_empty:
        type: string
      resolver_rule:
        $ref: '#/definitions/approval-engine-service_internal_domain.ResolverRule'
      step_order:
        type: integer
    type: object
  approval-engine-service_pkg_response.Envelope:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  approval-engine-service_pkg_response.Page:
    properties:
      items: {}
      meta:
        $ref: '#/definitions/approval-engine-service_pkg_response.PageMeta'
    type: object
  approval-engine-service_pkg_response.PageMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_handler.CreateRequestBody:
    properties:
      doc_type:
        example: purchase_order
        type: string
      payload:
        additionalProperties: {}
        type: object
      requester_id:
        example: SA01
        type: string
      resource_id:
        example: PO-2026-0912
        type: string
    type: object
  internal_handler.DecisionBody:
    properties:
      comment:
        type: string
      decision:
        enum:
        - approved
        - rejected
        example: approved
        type: string
      user_id:
        example: SS01
        type: string
    type: object
...
  internal_handler.PublishWorkflowBody:
    properties:
      app_id:
        example: assetmgmt
        type: string
      doc_type:
        example: purchase_order
        type: string
      name:
        example: PO Approval
        type: string
      steps:
        items:
          $ref: '#/definitions/internal_handler.StepInput'
        type: array
    type: object
  internal_handler.StepInput:
    properties:
      approval_mode:
        enum:
        - any
        - all
        example: any
        type: string
      condition:
        $ref: '#/definitions/approval-engine-service_internal_domain.Condition'
      conditions:
        items:
          $ref: '#/definitions/approval-engine-service_internal_domain.Condition'
        type: array
      logic:
        enum:
        - all
        - any
        example: all
        type: string
      name:
        example: Supervisor Approval
        type: string
      on_empty:
        enum:
        - fail
        - skip
        example: fail
        type: string
      resolver_rule:
        $ref: '#/definitions/approval-engine-service_internal_domain.ResolverRule'
    type: object
info:
  contact: {}
  description: |-
    Reusable approval workflow engine shared across Mayora internal systems (asset management first). A consuming app's requester_id/user_id are always checked against registered, active participants — the engine never trusts an identity blindly.
    Auth: every /requests route requires X-API-Key (a consuming application's server-to-server key — never expose it in a browser), and an application only ever sees and decides its own requests. /participants/import requires the operator X-Admin-Key. /portal/* serves the browser portal and can be switched off with PORTAL_ENABLED=false until SSO login is in place.
  title: Approval Engine API
  version: "1.0"
paths:
...
  /portal/workflows:
    get:
      description: Every version of every workflow (optionally filtered by app_id),
        for the management list view.
      parameters:
      - description: Filter by application ID
        in: query
        name: app_id
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_pkg_response.Page'
              type: object
      summary: List workflows
      tags:
      - workflows
    post:
      consumes:
      - application/json
      description: Always inserts a brand new version and atomically deactivates every
        other version of (app_id, doc_type) — same call whether the pair is new or
        already exists. In-flight requests keep running against the exact version
        they started with. Each step's resolver_rule/condition is validated immediately
        (unknown type, missing field, bad operator all fail here, not silently at
        resolve time). A step gates on either the single `condition` field or the
        compound `conditions`+`logic` fields ("all"/"any"), never both.
      parameters:
      - description: Workflow to publish
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.PublishWorkflowBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_internal_domain.WorkflowDefinition'
              type: object
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
      summary: Publish a workflow (create or edit)
      tags:
      - workflows
  /portal/workflows/{id}:
    get:
      parameters:
      - description: Workflow definition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_internal_domain.WorkflowDefinition'
              type: object
        "404":
          description: not found
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
      summary: Get a workflow version
      tags:
      - workflows
  /portal/workflows/{id}/deactivate:
    post:
      description: Stops the version from being used for new requests. Never a hard
        delete — past requests still reference this definition_id and must stay readable.
      parameters:
      - description: Workflow definition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
      summary: Deactivate a workflow version
      tags:
      - workflows
  /requests:
    post:
      consumes:
      - application/json
      description: 'Starts a new request against the calling application''s currently
        active workflow for doc_type. Idempotent on (app, doc_type, resource_id):
        calling it again with the same values returns the existing request instead
        of erroring, so a Spring Boot caller can safely retry on timeout. requester_id
        must be a registered, active participant.'
      parameters:
      - description: Request to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: already existed (idempotent replay)
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_internal_domain.ApprovalRequest'
              type: object
        "201":
          description: created
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_internal_domain.ApprovalRequest'
              type: object
        "400":
          description: validation error or unregistered requester
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
        "401":
          description: missing or invalid X-API-Key
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
      security:
      - ApiKeyAuth: []
      summary: Create an approval request
      tags:
      - requests
  /requests/{id}:
    get:
      description: Full request with materialized steps and assignment snapshots.
        An application can only read its own requests — another application's request
        id answers 404, so a key never reveals other apps' data.
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_internal_domain.ApprovalRequest'
              type: object
        "401":
          description: missing or invalid X-API-Key
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
      security:
      - ApiKeyAuth: []
      summary: Get a request
      tags:
      - requests
  /requests/{id}/decision:
    post:
      consumes:
      - application/json
      description: The calling application vouches for user_id (it has already authenticated
        that person in its own session); the engine then enforces that user_id is
        the person assigned to the currently active step and still an active participant.
        An application can only decide its own requests.
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Decision
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.DecisionBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/approval-engine-service_internal_domain.ApprovalRequest'
              type: object
        "400":
          description: not the assigned approver, already decided, or invalid decision
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
        "401":
          description: missing or invalid X-API-Key
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/approval-engine-service_pkg_response.Envelope'
      security:
      - ApiKeyAuth: []
      summary: Approve or reject the active step
      tags:
      - requests
...
securityDefinitions:
  AdminKeyAuth:
    in: header
    name: X-Admin-Key
    type: apiKey
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"