# internal/handler/workflow_handler.go
package handler

import (
	"context"
	"errors"
	"fmt"

	"github.com/gofiber/fiber/v2"
	"github.com/google/uuid"

	"approval-engine-service/internal/domain"
	"approval-engine-service/internal/service"
	"approval-engine-service/pkg/response"
)

type WorkflowStore interface {
	Publish(ctx context.Context, def *domain.WorkflowDefinition) error
	ListDefinitions(ctx context.Context, appID string, page, limit int) ([]domain.WorkflowDefinition, int, error)
	GetDefinitionByID(ctx context.Context, id string) (*domain.WorkflowDefinition, error)
	Deactivate(ctx context.Context, id string) error
}

// WorkflowHandler is the CRUD surface behind "add/edit/delete an approval
// flow without a code change". Editing publishes a new version rather than
// mutating one in place, so a request already in flight keeps running
// against the exact rules it started with (see WorkflowRepository.Publish).
type WorkflowHandler struct {
	workflows WorkflowStore
}

func NewWorkflowHandler(workflows WorkflowStore) *WorkflowHandler {
	return &WorkflowHandler{workflows: workflows}
}

// StepInput is one step in a PublishWorkflowBody. Set either Condition (the
// original single-condition shape) or Conditions+Logic (compound — "all" by
// default, or "any"), never both.
type StepInput struct {
	Name         string              `json:"name" example:"Supervisor Approval"`
	ResolverRule domain.ResolverRule `json:"resolver_rule"`
	Condition    *domain.Condition   `json:"condition,omitempty"`
	Conditions   []domain.Condition  `json:"conditions,omitempty"`
	Logic        string              `json:"logic,omitempty" example:"all" enums:"all,any"`
	ApprovalMode string              `json:"approval_mode" example:"any" enums:"any,all"`
	OnEmpty      string              `json:"on_empty" example:"fail" enums:"fail,skip"`
}

// PublishWorkflowBody is the payload for POST /workflows.
type PublishWorkflowBody struct {
	AppID   string      `json:"app_id" example:"assetmgmt"`
	DocType string      `json:"doc_type" example:"purchase_order"`
	Name    string      `json:"name" example:"PO Approval"`
	Steps   []StepInput `json:"steps"`
}

// Create publishes a new version of a workflow.
//
// @Summary Publish a workflow (create or edit)
// @Description Always inserts a brand new version and atomically deactivates every other version of (app_id, doc_type) — same call whether the pair is new or already exists. In-flight requests keep running against the exact version they started with. Each step's resolver_rule/condition is validated immediately (unknown type, missing field, bad operator all fail here, not silently at resolve time). A step gates on either the single `condition` field or the compound `conditions`+`logic` fields ("all"/"any"), never both.
// @Tags workflows
// @Accept json
// @Produce json
// @Param body body PublishWorkflowBody true "Workflow to publish"
// @Success 201 {object} response.Envelope{data=domain.WorkflowDefinition}
// @Failure 400 {object} response.Envelope "validation error"
// @Router /portal/workflows [post]
func (h *WorkflowHandler) Create(c *fiber.Ctx) error {
	var body PublishWorkflowBody
	if err := c.BodyParser(&body); err != nil {
		return response.Error(c, fiber.StatusBadRequest, "invalid request body")
	}
	if body.AppID == "" || body.DocType == "" || body.Name == "" {
		return response.Error(c, fiber.StatusBadRequest, "app_id, doc_type and name are required")
	}
	if len(body.Steps) == 0 {
		return response.Error(c, fiber.StatusBadRequest, "at least one step is required")
	}

	def := &domain.WorkflowDefinition{
		ID: uuid.NewString(), AppID: body.AppID, DocType: body.DocType, Name: body.Name,
	}
	for i, s := range body.Steps {
		mode := s.ApprovalMode
		if mode == "" {
			mode = domain.ModeAny
		}
		onEmpty := s.OnEmpty
		if onEmpty == "" {
			onEmpty = domain.OnEmptyFail
		}
		step := domain.WorkflowStep{
			ID: uuid.NewString(), StepOrder: i + 1, Name: s.Name,
			ResolverRule: s.ResolverRule, Condition: s.Condition, Conditions: s.Conditions, Logic: s.Logic,
			ApprovalMode: mode, OnEmpty: onEmpty,
		}
		if err := service.ValidateStep(step); err != nil {
			return response.Error(c, fiber.StatusBadRequest, fmt.Sprintf("step %d: %s", i+1, err.Error()))
		}
		def.Steps = append(def.Steps, step)
	}

	if err := h.workflows.Publish(c.Context(), def); err != nil {
		return response.Error(c, fiber.StatusBadRequest, err.Error())
	}
	for i := range def.Steps {
		def.Steps[i].DefinitionID = def.ID
	}
	return response.Success(c, fiber.StatusCreated, "workflow published", def)
}

// List returns every version of every workflow.
//
// @Summary List workflows
// @Description Every version of every workflow (optionally filtered by app_id), for the management list view.
// @Tags workflows
// @Produce json
// @Param app_id query string false "Filter by application ID"
// @Param page query int false "Page number (default 1)"
// @Param limit query int false "Items per page (default 20, max 100)"
// @Success 200 {object} response.Envelope{data=response.Page}
// @Router /portal/workflows [get]
func (h *WorkflowHandler) List(c *fiber.Ctx) error {
	page, limit := parsePagination(c)
	defs, total, err := h.workflows.ListDefinitions(c.Context(), c.Query("app_id"), page, limit)
	if err != nil {
		return response.Error(c, fiber.StatusInternalServerError, err.Error())
	}
	return response.Paginated(c, defs, page, limit, total)
}

// Get returns one workflow version with its steps.
//
// @Summary Get a workflow version
// @Tags workflows
// @Produce json
// @Param id path string true "Workflow definition ID"
// @Success 200 {object} response.Envelope{data=domain.WorkflowDefinition}
// @Failure 404 {object} response.Envelope "not found"
// @Router /portal/workflows/{id} [get]
func (h *WorkflowHandler) Get(c *fiber.Ctx) error {
	def, err := h.workflows.GetDefinitionByID(c.Context(), c.Params("id"))
	if err != nil {
		return response.Error(c, fiber.StatusInternalServerError, err.Error())
	}
	if def == nil {
		return response.Error(c, fiber.StatusNotFound, "workflow not found")
	}
	return response.Success(c, fiber.StatusOK, "", def)
}

// Deactivate is the "delete".
//
// @Summary Deactivate a workflow version
// @Description Stops the version from being used for new requests. Never a hard delete — past requests still reference this definition_id and must stay readable.
// @Tags workflows
// @Produce json
// @Param id path string true "Workflow definition ID"
// @Success 200 {object} response.Envelope
// @Failure 404 {object} response.Envelope "not found"
// @Router /portal/workflows/{id}/deactivate [post]
func (h *WorkflowHandler) Deactivate(c *fiber.Ctx) error {
	if err := h.workflows.Deactivate(c.Context(), c.Params("id")); err != nil {
		if errors.Is(err, domain.ErrWorkflowNotFound) {
			return response.Error(c, fiber.StatusNotFound, err.Error())
		}
		return response.Error(c, fiber.StatusInternalServerError, err.Error())
	}
	return response.Success(c, fiber.StatusOK, "workflow deactivated", nil)
}